+++ CHANGELOG.md
@@ -11,6 +11,6 @@
 - Added debug tracing of messages
 
 ## v1.2.0 - 19 January 2017
-- experimental workspace/packages endpoint
+- experimental workspace/xpackages endpoint
 - experimental workspace/symbols endpoint extension (symbol query)
 - textDocument/completion